fix(msteams): expose audit identity contract
