+++ extensions/msteams/src/channel.setup.ts
@@ -7,6 +7,7 @@ import {
   resolveMSTeamsDmPolicy,
   type ResolvedMSTeamsAccount,
 } from "./channel-config.js";
+import { collectMSTeamsSecurityFindings } from "./channel-security.js";
 import { MSTeamsChannelConfigSchema } from "./config-schema.js";
 import { msteamsSetupContract } from "./setup-core.js";
 import { msteamsSetupWizard } from "./setup-surface.js";
@@ -38,6 +39,7 @@ export const msteamsSetupPlugin: ChannelPlugin<ResolvedMSTeamsAccount> = {
   },
   security: {
     resolveDmPolicy: resolveMSTeamsDmPolicy,
+    collectWarnings: collectMSTeamsSecurityFindings,
   },
   setupWizard: msteamsSetupWizard,
   setupContract: msteamsSetupContract,
