+++ extensions/msteams/src/inbound.ts
@@ -1,6 +1,8 @@
 // Msteams plugin module implements inbound behavior.
 import { decodeHtmlEntities } from "openclaw/plugin-sdk/html-entity-runtime";
 
+export { normalizeMSTeamsConversationId } from "./ingress-identity.js";
+
 type MSTeamsQuoteInfo = {
   sender: string;
   body: string;
@@ -84,10 +86,6 @@ type MentionableActivity = {
     mentioned?: { id?: string };
   }> | null;
 };
-
-export function normalizeMSTeamsConversationId(raw: string): string {
-  return raw.split(";")[0] ?? raw;
-}
 
 export function extractMSTeamsConversationMessageId(raw: string): string | undefined {
   if (!raw) {
