+++ extensions/msteams/src/ingress-identity.ts
@@ -1,0 +1,121 @@
+// Msteams plugin module owns lightweight sender identity normalization and audit classification.
+import type {
+  IdentifierAuthentication,
+  StableChannelIngressIdentityParams,
+} from "openclaw/plugin-sdk/channel-ingress-runtime";
+import { normalizeOptionalLowercaseString } from "openclaw/plugin-sdk/string-coerce-runtime";
+
+const MSTEAMS_SENDER_NAME_KIND = "plugin:msteams-sender-name" as const;
+const MSTEAMS_CONVERSATION_ID_KIND = "plugin:msteams-conversation-id" as const;
+const MSTEAMS_AAD_OBJECT_ID = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
+const MSTEAMS_GROUP_CONVERSATION_ID = /^19:.+@thread\.(?:tacv2|skype|v2)$/i;
+
+function stripProviderPrefix(raw: string): string {
+  return raw.replace(/^(msteams|teams):/i, "");
+}
+
+export function normalizeMSTeamsUserInput(raw: string): string {
+  return stripProviderPrefix(raw)
+    .replace(/^(user|conversation):/i, "")
+    .trim();
+}
+
+export function normalizeMSTeamsConversationId(raw: string): string {
+  return raw.split(";")[0] ?? raw;
+}
+
+/** Detect supported prefixed and bare Bot Framework or Graph conversation IDs. */
+export function looksLikeMSTeamsConversationId(raw: string): boolean {
+  const trimmed = raw.trim();
+  if (!trimmed) {
+    return false;
+  }
+  if (/^conversation:/i.test(trimmed)) {
+    return true;
+  }
+  if (MSTEAMS_GROUP_CONVERSATION_ID.test(trimmed)) {
+    return true;
+  }
+  if (/^19:.+@unq\.gbl\.spaces$/i.test(trimmed)) {
+    return true;
+  }
+  if (/^a:1[A-Za-z0-9_-]+$/i.test(trimmed)) {
+    return true;
+  }
+  if (/^8:orgid:[A-Za-z0-9-]+$/i.test(trimmed)) {
+    return true;
+  }
+  return /@thread\b/i.test(trimmed);
+}
+
+function normalizeIngressValue(value?: string | null): string | null {
+  return normalizeOptionalLowercaseString(value) ?? null;
+}
+
+function normalizeSenderNameIngressValue(value?: string | null): string | null {
+  const normalized = normalizeIngressValue(value);
+  if (!normalized) {
+    return null;
+  }
+  // Conversation allowlist entries must never become spoofable display-name principals.
+  return looksLikeMSTeamsConversationId(normalizeMSTeamsConversationId(normalized))
+    ? null
+    : normalized;
+}
+
+function normalizeAllowlistConversationId(value?: string | null): string | null {
+  const trimmed = value?.trim();
+  // Microsoft Graph conversation IDs are opaque; case-folding would authorize a different chat.
+  return trimmed ? normalizeMSTeamsConversationId(trimmed) : null;
+}
+
+export const msteamsIngressIdentity = {
+  key: "sender-id",
+  // Bot Framework authenticates the connector and vouches for the activity, without this
+  // plugin independently proving exact ownership of every from.id representation.
+  authentication: "asserted",
+  normalize: normalizeIngressValue,
+  aliases: [
+    {
+      key: "sender-name",
+      kind: MSTEAMS_SENDER_NAME_KIND,
+      normalizeEntry: normalizeSenderNameIngressValue,
+      normalizeSubject: normalizeSenderNameIngressValue,
+      authentication: "mutable",
+    },
+    {
+      key: "conversation-id",
+      kind: MSTEAMS_CONVERSATION_ID_KIND,
+      authentication: "asserted",
+      normalizeEntry: normalizeAllowlistConversationId,
+      normalizeSubject: normalizeAllowlistConversationId,
+    },
+  ],
+  isWildcardEntry: (entry) => normalizeIngressValue(entry) === "*",
+  resolveEntryId: ({ entryIndex, fieldKey }) =>
+    `msteams-entry-${entryIndex + 1}:${
+      fieldKey === "sender-name"
+        ? "name"
+        : fieldKey === "conversation-id"
+          ? "conversation-id"
+          : "id"
+    }`,
+} satisfies StableChannelIngressIdentityParams;
+
+/** Classify authored DM allowlist entries without loading the Teams runtime. */
+export function classifyMSTeamsEntryAuthentication(
+  raw: string,
+): IdentifierAuthentication | undefined {
+  const unscoped = stripProviderPrefix(raw).trim();
+  const normalized = normalizeMSTeamsUserInput(raw);
+  if (
+    !normalized ||
+    normalized === "*" ||
+    /^conversation:/i.test(unscoped) ||
+    /^accessGroup:/i.test(normalized) ||
+    looksLikeMSTeamsConversationId(normalizeMSTeamsConversationId(normalized))
+  ) {
+    return undefined;
+  }
+  return MSTEAMS_AAD_OBJECT_ID.test(normalized) ? "asserted" : "mutable";
+}
